pyproject.toml
[build-system]  # https://docs.astral.sh/uv/concepts/projects/config/#build-systems
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]  # https://packaging.python.org/en/latest/specifications/pyproject-toml/
name = "raglite"
version = "1.0.0"
description = "A Python toolkit for Retrieval-Augmented Generation (RAG) with DuckDB, PostgreSQL, or SQLite."
readme = "README.md"
authors = [
  { name = "Laurent Sorber", email = "laurent@superlinear.eu" },
]
requires-python = ">=3.10,<4.0"
dependencies = [
  # Configuration:
  "platformdirs (>=4.0.0)",
  # Markdown conversion:
  "pdftext (>=0.4.1)",
  "scikit-learn (>=1.4.2)",
  # Sentence and chunk splitting:
  "markdown-it-py (>=3.0.0)",
  "numpy (>=1.26.4,<2.0.0)",
  "scipy (>=1.11.2,!=1.15.0.*,!=1.15.1,!=1.15.2)",
  "wtpsplit-lite (>=0.1.0)",
  # Large Language Models:
  "huggingface-hub (>=0.31.2)",
  "litellm (>=1.60.2)",
  "pydantic (>=2.7.0)",
  # Storage:
  "duckdb (>=1.1.3)",
  "duckdb-engine (>=0.16.0)",
  "pg8000 (>=1.31.2)",
  "sqlmodel-slim (>=0.0.21)",
  "sqlite-vec (>=0.1.0)",
  "pynndescent (>=0.5.12)",
  # Progress:
  "tqdm (>=4.66.0)",
  # CLI:
  "typer (>=0.15.1)",
  # Model Context Protocol:
  "fastmcp (>=2.0.0)",
  # Utilities:
  "packaging (>=23.0)",
]

[project.scripts]  # https://docs.astral.sh/uv/concepts/projects/config/#command-line-interfaces
raglite = "raglite._cli:cli"

[project.urls]  # https://packaging.python.org/en/latest/specifications/well-known-project-urls/#well-known-labels
homepage = "https://github.com/superlinear-ai/raglite"
source = "https://github.com/superlinear-ai/raglite"
changelog = "https://github.com/superlinear-ai/raglite/blob/main/CHANGELOG.md"
releasenotes = "https://github.com/superlinear-ai/raglite/releases"
documentation = "https://github.com/superlinear-ai/raglite"
issues = "https://github.com/superlinear-ai/raglite/issues"

[dependency-groups]  # https://docs.astral.sh/uv/concepts/projects/dependencies/#development-dependencies
dev = [
  "commitizen (>=4.3.0)",
  "coverage[toml] (>=7.6.10)",
  "ipykernel (>=6.29.4)",
  "ipython (>=8.18.0)",
  "ipywidgets (>=8.1.2)",
  "mypy (>=1.14.1)",
  "pdoc (>=15.0.1)",
  "poethepoet (>=0.32.1)",
  "pre-commit (>=4.0.1)",
  "pytest (>=8.3.4)",
  "pytest-mock (>=3.14.0)",
  "pytest-xdist (>=3.6.1)",
  "ruff (>=0.10.0)",
  "typeguard (>=4.4.1)",
]

[project.optional-dependencies]  # https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies
# Frontend:
chainlit = ["chainlit (>=2.0.0)"]
# Large Language Models:
llama-cpp-python = ["llama-cpp-python (>=0.3.9)"]
# Markdown conversion:
pandoc = ["pypandoc-binary (>=1.13)"]
# Reranking:
rerank = [
  "langdetect (>=1.0.9)",
  "rerankers[api,flashrank] (>=0.10.0)",
]
gpu = [
  "torch>=2.0.0",              # GPU detection and management
  "llama-cpp-python>=0.2.0",   # GPU-enabled LlamaCpp
]
# Evaluation:
ragas = ["pandas (>=2.1.1)", "ragas (>=0.1.12)"]
# Benchmarking:
bench = [
  "faiss-cpu (>=1.11.0)",
  "ir_datasets (>=0.5.10)",
  "ir_measures (>=0.3.7)",
  "llama-index (>=0.12.39)",
  "llama-index-vector-stores-faiss (>=0.4.0)",
  "openai (>=1.75.0)",
  "pandas (>=2.1.1)",
  "python-slugify (>=8.0.4)",
]

[tool.commitizen]  # https://commitizen-tools.github.io/commitizen/config/
bump_message = "bump: v$current_version → v$new_version"
tag_format = "v$version"
update_changelog_on_bump = true
version_provider = "uv"

[tool.coverage.report]  # https://coverage.readthedocs.io/en/latest/config.html#report
fail_under = 50
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.run]  # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
command_line = "--module pytest"
data_file = "reports/.coverage"
source = ["src"]

[tool.coverage.xml]  # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"

[tool.mypy]  # https://mypy.readthedocs.io/en/latest/config_file.html
junit_xml = "reports/mypy.xml"
strict = true
disallow_subclassing_any = false
disallow_untyped_decorators = false
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[tool.pytest.ini_options]  # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--color=yes --doctest-modules --ignore=src/raglite/_chainlit.py --exitfirst --failed-first --strict-config --strict-markers --verbosity=2 --junitxml=reports/pytest.xml"
filterwarnings = ["error", "ignore::DeprecationWarning", "ignore::pytest.PytestUnraisableExceptionWarning"]
markers = ["slow: mark test as slow", "benchmark: mark test as benchmark"]
testpaths = ["src", "tests"]
xfail_strict = true

[tool.ruff]  # https://docs.astral.sh/ruff/settings/
fix = true
line-length = 100
src = ["src", "tests"]
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["A", "ASYNC", "B", "BLE", "C4", "C90", "D", "DTZ", "E", "EM", "ERA", "F", "FBT", "FLY", "FURB", "G", "I", "ICN", "INP", "INT", "ISC", "LOG", "N", "NPY", "PERF", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "Q", "RET", "RSE", "RUF", "S", "SIM", "SLF", "SLOT", "T10", "T20", "TCH", "TID", "TRY", "UP", "W", "YTT"]
ignore = ["D203", "D213", "E501", "RET504", "RUF002", "RUF022", "S101", "S307", "TC004"]
unfixable = ["ERA001", "F401", "F841", "T201", "T203"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.poe.executor]  # https://github.com/nat-n/poethepoet
type = "simple"

[tool.poe.tasks]

  [tool.poe.tasks.docs]
  help = "Generate this package's docs"
  cmd = """
    pdoc
      --docformat $docformat
      --output-directory $outputdirectory
      raglite
    """

    [[tool.poe.tasks.docs.args]]
    help = "The docstring style (default: numpy)"
    name = "docformat"
    options = ["--docformat"]
    default = "numpy"

    [[tool.poe.tasks.docs.args]]
    help = "The output directory (default: docs)"
    name = "outputdirectory"
    options = ["--output-directory"]
    default = "docs"

  [tool.poe.tasks.lint]
  help = "Lint this package"
  cmd = """
    pre-commit run
      --all-files
      --color always
    """

  [tool.poe.tasks.test]
  help = "Test this package"

    [[tool.poe.tasks.test.sequence]]
    cmd = "coverage run"

    [[tool.poe.tasks.test.sequence]]
    cmd = "coverage report"

    [[tool.poe.tasks.test.sequence]]
    cmd = "coverage xml"
